make steffen CFColor fix also work on older gcc versions used in grid-sites (there is no 'int pow(int,int)' in gcc-4.1.2)

+++ AMEGIC++/Amplitude/CFColor.C
@@ -565,7 +565,7 @@ string CFColor::BuildTChain(vector<string>  string_list, bool & valid)
   
   size_t length = string_list.size();
   //approximate length faculty
-  size_t nperm = pow(length,length-1);
+  size_t nperm = pow((double)length,(int)(length-1));
   
   //generate the traces  
   for(;;) {
